perf(compiler, mapper): collect all sequences of KSP immediately and don't allow external interfaces
With this is not double processing the same declaration or processing outside the expected scope time because a sequence is iterated all the time a terminal operator is used. For example, before this change in logs, I could see more than one processed function X when only one X function existed. The same happens for all sequences of ClassData